upgrade to pre-release
